Configure logger
Why these changes are being introduced:

As a library that is installed and imported, it is important that
logging is setup in a way that aligns with the calling application.

How this addresses that need:
* Function configure_logger returns a logger instance that is
configured properly and uses an optional env var TDA_LOG_LEVEL
to set the level for only this logger.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-415