Dynamic library version from package
Why these changes are being introduced:

Formerly, the python library version number was set via the
pyproject.toml file.  This worked, as the actual package itself
could provide the version by reading the installed package
metadata.  However, some best practices discourage this, as it
relies on the packing and metadata reading.

We may choose to revisit this approach as we get farther along,
particularly if installing via a Github tag or release is
important.

How this addresses that need:
* Version is set in timdex_dataset_api.__init__.__version__,
allowing it to be read directly from the library
* Packaging via pyproject.toml uses this version number via a
"dynamic" metadata field

Side effects of this change:
* None at this time

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-414