Add dataset writing functionality
Why these changes are being introduced:

A primary use case for the library will be Transmogrifier writing new
records to the parquet dataset.  This library is intended to make
that work simple for Transmogrifier, where all it needs to do is
yield DatasetRecords (imported from this lib) to the write method.

How this addresses that need:
* Adds new entrypoint write() method
* Includes helper methods to batch records yielded to write method
* Adds DatasetRecord class that is designed to encapsulat each
record (row) that will get written to the dataset.

Side effects of this change:
* Library supports writing to local or remote dataset

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-415