Validate DatasetRecord during serialization
Why these changes are being introduced:

When a DatasetRecord is serialized for writing, it contains the dataset
partitions that the record will fall under.  We do not want any records
getting written to the dataset without all required partitions.

Additionally, we may want to add future validations to the record, e.g.
ensuring the transformed record is valid JSON.  This provides a place
for that.

How this addresses that need:
* Adds new DatasetRecord.validate() method, and applies method during
to_dict() serialization.

Side effects of this change:
* Records cannot be written to dataset without all required partitions

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-415