Reorder run_id and action partitions
Why these changes are being introduced:

Discussions around retrieving subsets of records from the dataset suggested
that it would be beneficial to have the run_id partition before the
action partition.  This will allow using a prefix approach of partition
names and values when loading a dataset, that will pinpoint a particular
run even before the dataset is fully loaded.

This ordering was originally proposed in the engineering plan for this
library, but it switched somewhere along the way; so moving back to
agreed upon ordering.

How this addresses that need:
* Moves run_id before action in ordered partition columns

Side effects of this change:
* Omitting the action partition, and using everything until the
run_id partition, is sufficient for getting all records from a run.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-424