Replace static fixture of local dataset with temp, dynamic dataset
Why these changes are being introduced:
* This avoids the creation of artifacts in the 'tests/fixtures/' directory
that become outdated if / when partitioning scheme changes. This also
updates the 'local_dataset' fixture to use the 'tmp_path' fixture from
pytest, allowing each test that uses the 'local_dataset' fixture
to run the test in a clean and isolated environment.

How this addresses that need:
* Update "local_dataset" fixtures
* Update unit tests

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-440