Remove option to load dataset using partition prefix
Why these changes are being introduced:
* When a partition prefix is constructed for the dataset with year, month, and day (or a combination of them),
and appended to the 'source' arg in ds.dataset, the resulting dataset will
contain 'None' values for any partition columns used in the prefix. It is then
problematic with the "post-load" filtering step, which again attempts
to filter on those partition columns. It was believed that would be
somewhat inconsequential, but additional testing revealed this was not the case.
These changes simplify the TIMDEXDataset.load method by instead
relying on PyArrow's efficient dataset discovery and reading processes.
For more details, see comment on PR #31:
https://github.com/MITLibraries/timdex-dataset-api/pull/31#pullrequestreview-2517764044.

How this addresses that need:
* Remove '_get_partition_prefixes' private method
* Update '_parse_date_filters' to raise TypeError
* Update unit tests

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-425