Create typed DatasetFilters object
Why these changes are being introduced:

Currently, both loading and filtering the dataset resulted in a large
number of keyword arguments for each method to aligned with dataset
columns.  Moving into read methods, which will also support filtering,
this was a substantial amount of duplication and could be error prone
over time.

How this addresses that need:
* Creates a typed dictionary DatasetFilters that includes all columns
or partitions that we can use when filtering the dataset
* Each method that can filter the dataset accepts kwargs, but they are
typed to this typed dictionary

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-417