Add baseline read methods to TIMDEXDataset
Why these changes are being introduced:

A primary responsibility of the TIMDEXDataset class is to provide
performant and memory efficient reading of a dataset.  It is anticipated
that additional read methods may be required, for specific or niche
situations, but some simple baseline ones are needed at this time.

How this addresses that need:
* Adds methods for reading pyarrow batches, pandas dataframes, and python
dictionaries from a dataset.

Side effects of this change:
* Applications like timdex lambdas or TIM can now read records from dataset

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-417