Simplify logging setup
Why these changes are being introduced:

As a library that is installed by other applications, it is beneficial to
have this library logging follow the levels and handlers set by the calling
context.  That said, it can also be helpful to set a level like INFO for the
calling context, but DEBUG for this library.

What we definitely want to avoid is this library setting up a handler and
outputting logs, only to have them duplicated in the calling context.

How this addresses that need:
* Remove setting up of handlers; rely on calling context for handlers

Side effects of this change:
* For local development logging statements will not show up without
some kind of manual logging.basicConfig() call or handler setup.
* For calling contexts Transmogrifier, pipeline lambdas, and TIM to show this
library's DEBUG logs, they will need to adjust their configure_logger() functions
to avoid filtering out loggers from different namespaces.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-427