Refactor to configuration object
Why these changes are being introduced:

With the addition of two read configurations that would be passed
around beteween multiple methods, the dataset module tipped
to where a centralized configuration object would be helpful.

Additionally, we have learned that per-operation configurations are
rare, and much more likely to be set once during TIMDEXDataset
init, or even as env vars for the duration of the library import.

How this addresses that need:

Creates a dataclass TIMDEXDatasetConfig that is passed to
TIMDEXDataset on init.  This class provides a typed object, with sensible
defaults, that are shared throughout all read and write methods.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-468