Support filtering for current_records
Why these changes are being introduced:

Unexpected behavior was possible when using load(current_records=True)
and then applying additional filtering to the dataset before reading.
In short, a non-current record could be yielded if filtering removed
the truly current version of the record.  This happened because the
reverse chronological marking of "seen" records would not "see" this
record and happily yield an older version.

How this addresses that need:

When load(current_records=True) is used, a clone of the dataset is
saved to the TIMDEXDataset object before any additional filtering
is applied.  This dataset is just metadata, not expensive to store.

Then, during any read methods, this dataset is used to provide an
exhaustive and ordered list of timdex_record_ids.  Even if a record
has been filtered out by the read method (e.g. limiting records to only
action="index"), this secondary list of timdex_record_ids is used
as the authoritative list of "seen" timdex_record_ids.

There is a bit of network overhead to this parallel batch reading,
but fairly minimal as we are only retrieving the 'timdex_record_id';
perhaps 1-2mb of IO per millions of records.

Side effects of this change:
* Applications like TIM that will likely use this new functionality
to yield only "current" records can do so confidently, and optionally
with additional filtering, knowing they will only encounter current
versions of a record from the dataset.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-497