DatasetRecord supports explicit run_timestamp
Why these changes are being introduced:

After it was discovered that a single ETL run (run_id) has multiple, different
run_timestamps in the dataset, it was clear we needed a way to pass an explicit
run_timestamp for writing instead of relying on a run_timestamp minted as part
of the TIMDEXDataset.write() method.

How this addresses that need:

DatasetRecord was given an optional run_timestamp property, that defaults
to using run_date and producing a full ISO timestamp from that.

Side effects of this change:
* Applications like Transmogrifier can pass an explicit run_timestamp
when writing to the dataset, ensuring that even multiple invocations
of Transmogrifier + TIMDEXDataset.write() can share the same timestamp
for an ETL run.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-509