Create TIMDEXDatasetMetadata class
Why these changes are being introduced:

As outlined in more detail in TIMX-506, this class is a new, alternate approach
to providing high level metadata about the dataset to better support bulk
operations.

This class is designed to remove the need for TIMDEXRunManager, while also providing
more helpful and granular data about the dataset.  Ultimately this will reduce code
complexity by relying less on implicit data contracts about the datasaet and more
on verifiable, data-in-hand about the dataset that can be used for bulk operations
(e.g. yielding current records).

How this addresses that need:
* New class TIMDEXDatasetMetadata is created
  * uses DuckDB to quickly crawl the dataset, generating an in-memory or on-disk
    database of record metadata
* This class is not yet wired into any TIMDEXDataset operations, this will happen
in future work

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-506