Current records functionality uses TIMDEXDatasetMetadata
Why these changes are being introduced:

Previously, when asked to yield only current records the class
TIMDEXRunManager was used to crawl the dataset and produce a dataframe of
ETL runs that informed a second iterator to consult during the yielding
of records.  This class was superceded by TIMDEXDatasetMetadata which
can provide the same information, but also more granular record level
metadata which simplifies the yielding of current records.

How this addresses that need:

The new TIMDEXDatasetMetadata class is used to provide a mapping
of timdex_record_id to run_id that allows, during the yielding of records,
to only yield a record if it's associated with the correct run_id.

The net effect here is a simpler, and ultimately more accureate, way to
yield current records while also utilizing a class that provides more
functionality in other contexts.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-507