Add 003 row group size data migration
Why these changes are being introduced:

It was discovered that some parquet files, mostly those before
2025-06-27, have potentially a single, very large row group instead
of maximum rows per group of 1k.

This has been addressed in recent work, resulting in no files after
2025-06-28 having a single row group if total rows exceed 1k.

How this addresses that need:

Migration script that re-packs row groups (rewrites) when a single
row group is found for a number of rows that should be split up.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-512