Add MinIO S3 support
Why these changes are being introduced:

As we prepare for expanding the role of dataset metadata,
which may be stored as a static file in S3 next to the dataset,
it will be helpful to test and develop against a realistic S3
backdrop but do so locally.

How this addresses that need:

Configures pyarrow and DuckDB to create appropriate connections
when a 'MINIO_S3_ENDPOINT_URL' env var is present.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-515