Begin rebuild of TIMDEXDatasetMetadata
Why these changes are being introduced:

The current overarching work is to support the creation and
reading of a static metadata database file and append deltas.  To
get there, very little of the original TIMDEXDatasetMetadata class
is needed or wanted.

This commit begins the process of rebuilding TIMDEXDatasetMetadata,
oriented around managing a static metadata database file, and providing
a readonly projection over that and append delta paqruet files.

How this addresses that need:

TIMDEXDatasetMetadata is almost completely rebuilt, with the first
functionality being the creation of the static metadata file by
scanning the ETL records.  Then, the ability to remotely attach in
readonly mode to this metadata database file for reading.

Note: these changes are breaking.  TIMDEXDataset cannot provide
"current" records and many unit tests are broken.  This will be
addressed in future commits as we build this class back up
with new functionality.

Side effects of this change:
* TIMDEXDataset cannot provide current records
* Unit tests are either temporarily skipped or failing

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-530