Begin rebuilding of data and metadata tests
Why these changes are being introduced:

With the big changes to TIMDEXMetadataDataset comes the need to virtually
rewrite the test suite for that class.

The changes too TIMDEXMetadataDataset are also influencing tests for
TIMDEXDataset, both how its loaded and tested for 'current' record reading.

How this addresses that need:

This begins with some basic tests around the loading, creating, and attaching
of a static database file for TIMDEXMetadataDataset.  Future tests will
more fully exercise the final views and tables created.

This commit also *temporarily* skips a bunch of tests for TIMDEXDataset
that will not pass until the ability to limit to 'current' records
is reinsated with the updated TIMDEXMetadataDataset.

Side effects of this change:
* Test suite passes, but multiple tests are temporarily skipped.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-530