Add append delta filename to metadata.append_deltas view
Why these changes are being introduced:
* The method for merging append deltas into the static metadata database file
needs the filenames of append deltas to easily identify which files
to delete (once merged). Prior to this change, the append deltas view
only had a 'filename' column, which referred to the path or S3 URI for
the TIMDEXDataset parquet file.

How this addresses that need:
* Set filename='append_delta_filename' when creating metadata.append_deltas view
* Explicitly select metadata column names when creating metadata.records view

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-528