Bump version to v3.0
Why these changes are being introduced:

The shift to elevating dataset metadata as an expected and required asset,
and utilizing DuckDB context's for metadata and data retrieval is a
sizable and breaking change.  Feels like this warrants a major version
bump.

How this addresses that need:
* Bumps internal library version to 3.0

Side effects of this change:
* `make update` commands from TIMDEX components will pickup this new version

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-537