current_records view relies on temp table
Why these changes are being introduced:

The current_records view has always been a performance and resource
bottleneck.  Moving to metadata and SQL has helped, but there was a little
kink left in relation to using that metadata for data retreival.

We often would materialize a query to a pandas dataframe for to use
to drive data retrieval.  In that moment, we do not benefit from having
current_records be a view, when we're going to materialize the data
anyhow.

How this addresses that need:

Utilizing a DuckDB temp table, we take a small performance hit on
TIMDEXDatasetMetadata load, but then have near instant current_records
queries thereafter.

Additionally, we remove ordering in the metadata query for data
retrieval and perform this in-memory with the pandas dataframe.  Often
this may be quite small, but even if large, it's more efficient here
and already in python memory.

This will also set the stage for performing just-in-time metadata
queries as chunks before data retrieval, versus pulling all metadata
rows in one query and then chunking that in memory.

Side effects of this change:
* Quicker metadata queries, small performance hit on load.  Appears
similarly memory intensive.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-543