Add LIMIT clause to read methods
Why these changes are being introduced:

Sometimes it can be helpful to limit the results from a read method.

How this addresses that need:
Adds optional limit= arg to all read methods which is passed along to
the metadata query.  By limiting the metadata results, we limit the
data records retrieved.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-543