Pin python duckdb to 1.3.2
Why these changes are being introduced:

The duckdb python client recently released 1.4.0.  It has some exciting
new features that we can utilize, but it introduces some minor changes that conflict
with another 3rd party library, duckdb_engine, that we use for creating a
SQLAlchemy engine object from a DuckDB connection.

While upgrading to 1.4.0 will be a net positive for us, we are utilizing
this library heavily in other contexts.

How this addresses that need:

Pins our python duckdb client to previous release 1.3.2, which maintains
current functionality. A new ticket will be created to work through the
1.4.0 upgrade.

Side effects of this change:
* None

Relevant ticket(s):
* None