Pin duckdb to 1.4.2-dev27
Why these changes are being introduced:

Previously, we had to pin our duckdb version to <1.4 as the 1.4 release had a couple of temporary
breaking changes.  Once 1.4.1 was released we tried unpinning, but found a new breaking change.
This was again determined temporary after some duckdb issues discussions, indicating that a 1.4.2
release will fix this bug.

In both cases they were bugs with the release.  It is believed that the 1.4 release is a big one
for duckdb, so a bit more churn that normal.

That said, development is rapid on the TIMDEX ETL pipeline at the moment, and this library, and
continuing to work with the newest version of duckdb is desirable.  A 1.4.2 release will likely
be soon, but we don't know when.

How this addresses that need:

Updates duckdb pinned version to 1.4.2.dev27, which is confirmed to work for us.  Once 1.4.2 is
released, we can remove this pin.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-559